接入etcd

+++ model/payRecord.go
@@ -6,7 +6,7 @@ type PayRecord struct {
 	AppId      string  `gorm:"app_id"`
 	ProductId  int64   `gorm:"product_id"`
 	UserId     int64   `gorm:"user_id"`
-	PayAmount  float64 `gorm:"pay_amount"`
+	PayAmount  float32 `gorm:"pay_amount"`
 	PayStatus  string  `gorm:"pay_status"`
 	PayMethod  string  `gorm:"pay_method"`
 	CreateTime int64   `gorm:"create_time"`
